requirements.txt
adafruit-circuitpython-mcp3xxx
ansible-vault
paho-mqtt
pushnotifier
python-crontab
RPi.GPIO
requests
setuptools